enhance: user presets are edit-panel rows; 32-row mask (ai2)
The enhance-presets user graphs were appended after the eight built-ins,
but the mask was a guint8 and the panel stopped at eight cards, so no
user preset could ever be toggled or tuned.

- The panel lists every preset: the built-ins, then the user's own in
  Preferences order, up to GGAZE_ENHANCE_MAX_PRESETS (32) rows --
  GGAZE_ENHANCE_MAX_USER_PRESETS (24) user presets; the rest are ignored.
- The mask is a guint32 end to end (EditSnapshot, saved state, undo,
  title, render chain, export, thumbnails); strength arrays are sized to
  the cap.
- j/k reach every row and scroll it (with its slider) into view once it
  is laid out; Enter and h/l act on the selected row; digits 1-8 stay the
  first eight rows, rows past them show no digit.
- A Preferences change of the user presets carries the edit to the rows
  its presets have now (enhancer_presets_map: same name+graph, else same
  name = graph edited, else same graph = renamed); a removed preset's
  state is dropped. The undo history is remapped (edit_history_remap),
  not cleared; the saved state stays saved only if it still renders the
  same (enhancer_chain_key); the picture re-renders only when the chain
  changed; an open panel is rebuilt in place.
- A digit can no longer toggle a bit with no preset behind it.

+++ src/enhancer-gegl.h
@@ -43,7 +43,7 @@ GeglBuffer *enhancer_apply(GeglBuffer *p_in, const EnhancerPreset *p_preset,
  * size), which is what lets the crop tool lay out its rectangle before the
  * preview has rendered. */
 GeglBuffer *enhancer_apply_chain(GeglBuffer *p_in, const GPtrArray *p_presets,
-                                 guint8 u_mask, const Transform *p_xf,
+                                 guint32 u_mask, const Transform *p_xf,
                                  GError **p_err);
 
 /* Export the enhanced buffer to a file. The saver is chosen from p_out's
@@ -65,7 +65,7 @@ gboolean enhancer_export(GeglBuffer *p_in, const EnhancerPreset *p_preset,
 /* Export p_in with the enabled-preset chain (u_mask) and the transform p_xf
  * (nullable) composed, to p_out. */
 gboolean enhancer_export_chain(GeglBuffer *p_in, const GPtrArray *p_presets,
-                               guint8 u_mask, const Transform *p_xf,
+                               guint32 u_mask, const Transform *p_xf,
                                GFile *p_out, GError **p_err);
 
 /* Async export: load p_src, apply the chain + transform, save to p_out --
@@ -74,7 +74,7 @@ gboolean enhancer_export_chain(GeglBuffer *p_in, const GPtrArray *p_presets,
  * (AGENTS.md: decode runs in GTask threads). p_presets and p_xf (nullable)
  * are snapshotted. Finish returns TRUE on a real write. */
 void     enhancer_export_chain_async(GFile *p_src, const GPtrArray *p_presets,
-                                     guint8 u_mask, const Transform *p_xf,
+                                     guint32 u_mask, const Transform *p_xf,
                                      GFile *p_out, GCancellable *p_cancel,
                                      GAsyncReadyCallback p_cb, gpointer p_data);
 gboolean enhancer_export_chain_finish(GAsyncResult *p_res, GError **p_err);
@@ -129,7 +129,7 @@ GdkTexture *enhancer_buffer_to_texture(GeglBuffer *p_buf, GError **p_err);
  * newer apply superseding this one) must still check that on its own
  * before using the finished result. */
 void enhancer_apply_chain_async(GFile *p_file, const GPtrArray *p_presets,
-                                guint8 u_mask, const Transform *p_xf,
+                                guint32 u_mask, const Transform *p_xf,
                                 GCancellable       *p_cancel,
                                 GAsyncReadyCallback p_cb, gpointer p_data);
